lil bit of refactoring
