+++ core/src/library/models/mod.rs
@@ -1,3 +1,4 @@
+pub mod file;
 pub mod library_item;
 pub mod library_item_version;
 pub mod library_item_version_type;
